backend/requirements.txt
fastapi
uvicorn
yt-dlp
openai-whisper
python-multipart
ffmpeg-python
torch
numpy
firebase-admin
